Refactor insertion function
