--- src/koeda/random_insertion.py
@@ -1,27 +1,0 @@
-import random
-
-from .utils import get_synonyms
-
-
-def add_word(new_words):
-    synonyms = []
-    counter = 0
-    while len(synonyms) < 1:
-        random_word = new_words[random.randint(0, len(new_words) - 1)]
-        synonyms = get_synonyms(random_word)
-        counter += 1
-        if counter >= 10:
-            return
-    random_synonym = synonyms[0]
-    random_idx = random.randint(0, len(new_words) - 1)
-    new_words.insert(random_idx, random_synonym)
-
-def random_insertion(words, n):
-    new_words = words.copy()
-    for _ in range(n):
-        add_word(new_words)
-    return new_words
-
-
-
-
